+++ gazebo/gui/model/JointMaker.hh
@@ -47,6 +47,7 @@ namespace gazebo
 {
   namespace gui
   {
+    class JointCreationDialog;
     class JointData;
     class JointInspector;
 
@@ -109,13 +110,14 @@ namespace gazebo
       /// \brief Create a joint with parent and child.
       /// \param[in] _parent Parent of the joint.
       /// \param[in] _child Child of the joint.
-      /// \return joint data.
+      /// \return Joint data.
       public: JointData *CreateJoint(rendering::VisualPtr _parent,
           rendering::VisualPtr _child);
 
       /// \brief Helper method to create hotspot visual for mouse interaction.
       /// \param[in] _joint Joint data used for creating the hotspot
-      public: void CreateHotSpot(JointData *_joint);
+      /// \return Joint id, empty if hotspot creation failed.
+      public: std::string CreateHotSpot(JointData *_joint);
 
       /// \brief Update callback on PreRender.
       public: void Update();
@@ -139,11 +141,13 @@ namespace gazebo
       /// \brief Generate SDF for all joints.
       public: void GenerateSDF();
 
-      /// \brief Generate SDF for all joints.
+      /// \brief Get model SDF element containing all joints.
+      /// \return Pointer to SDF element.
       public: sdf::ElementPtr GetSDF() const;
 
       /// \brief Get the axis count for joint type.
       /// \param[in] _type Type of joint.
+      /// \return Axis count.
       public: static unsigned int GetJointAxisCount(
           JointMaker::JointType _type);
 
@@ -163,15 +167,16 @@ namespace gazebo
       public: static std::string GetJointMaterial(const std::string &_type);
 
       /// \brief Get state
-      /// \return State of JointType if joint creation is in process, otherwise
-      /// JOINT_NONE
+      /// \return Current state of the joint maker. If mouse is enabled to
+      /// create a new joint, it returns the type of joint. Otherwise, it
+      /// returns JOINT_NONE.
       public: JointMaker::JointType GetState() const;
 
       /// \brief Stop the process of adding joint to the model.
       public: void Stop();
 
       /// \brief Get the number of joints added.
-      /// return Number of joints.
+      /// \return Number of joints.
       public: unsigned int GetJointCount();
 
       /// \brief Create a joint from SDF. This is mainly used when editing
@@ -206,6 +211,56 @@ namespace gazebo
       /// \brief Get the list of links.
       /// \return The list of links, with the link scoped name and leaf name.
       public: std::map<std::string, std::string> LinkList() const;
+
+      /// \brief A new type for the joint being created has been chosen.
+      /// To be used by other classes.
+      /// \param[in] _typeInt Integer corresponding to joint type enum.
+      public slots: void OnType(const int _typeInt);
+
+      /// \brief A new axis for the joint being created has been chosen.
+      /// To be used by other classes.
+      /// \param[in] _axis Axis which was changed, either "axis1" or "axis2".
+      /// \param[in] _value New value for the axis
+      public slots: void SetAxis(const QString &_axis,
+          const ignition::math::Vector3d &_value);
+
+      /// \brief A new joint pose for the joint being created has been chosen.
+      /// To be used by other classes.
+      /// \param[in] _pose New joint pose.
+      public: void SetJointPose(const ignition::math::Pose3d &_pose);
+
+      /// \brief A new parent link for the joint being created has been chosen.
+      /// To be used by other classes.
+      /// \sa SetParentLink(rendering::VisualPtr _parentLink)
+      /// \param[in] _name Link name, either the leaf or scoped.
+      public: void SetParentLink(const std::string &_name);
+
+      /// \brief A new child link for the joint being created has been chosen.
+      /// To be used by other classes.
+      /// \sa SetChildLink(rendering::VisualPtr _childLink)
+      /// \param[in] _name Link name, either the leaf or scoped.
+      public: void SetChildLink(const std::string &_name);
+
+      /// \brief A new relative pose for the child link of the joint being
+      /// created hass been chosen. The pose is expressed in the parent link
+      /// frame.
+      /// \param[in] _pose New pose.
+      /// \param[in] _reset Set to true to reset the relative pose to the
+      /// original one.
+      public: void SetLinksRelativePose(
+          const ignition::math::Pose3d &_pose, const bool _reset);
+
+      /// \brief Align the parent and child links of the joint being created.
+      /// \param[in] _childToParent True to align the child to the parent,
+      /// false to align the parent to the child.
+      /// \param[in] _axis Axis of alignment (x/y/z)
+      /// \param[in] _mode Alignment mode (min/center/max)
+      public: void AlignLinks(
+          const bool _childToParent, const std::string &_axis,
+          const std::string &_mode);
+
+      /// \brief Finalize joint creation.
+      public: void FinalizeCreation();
 
       /// \brief Mouse event filter callback when mouse button is pressed.
       /// \param[in] _event The mouse event.
@@ -272,7 +327,7 @@ namespace gazebo
       private: void OnLinkInserted(const std::string &_linkName);
 
       /// \brief Remove a link from the list.
-      /// \param[in] _linkId Unique link identifying name.
+      /// \param[in] _linkName Unique link identifying name.
       private: void OnLinkRemoved(const std::string &_linkName);
 
       /// \brief Create a joint line.
@@ -282,15 +337,30 @@ namespace gazebo
       private: JointData *CreateJointLine(const std::string &_name,
           rendering::VisualPtr _parent);
 
+      /// \brief Get a link's visual from the link's name.
+      /// \param[in] _name Link name, scoped or not.
+      /// \return Pointer to link visual.
+      private: rendering::VisualPtr LinkVisualFromName(
+          const std::string &_name);
+
       /// \brief Set a new parent link for the joint being created.
+      /// \sa SetParentLink(const std::string &_name);
       /// \param[in] _parentLink Pointer to the link visual.
       /// \return True if successfully set new parent.
       private: bool SetParentLink(rendering::VisualPtr _parentLink);
 
       /// \brief Set a new child link for the joint being created.
+      /// \sa SetChildLink(const std::string &_name);
       /// \param[in] _childLink Pointer to the link visual.
       /// \return True if successfully set new child.
       private: bool SetChildLink(rendering::VisualPtr _childLink);
+
+      /// \brief Highlight link visuals which have been moved while creating
+      /// a new joint.
+      /// \param[in] _vis Visual to be highlighted.
+      /// \param[in] _moved Whether it moved or not.
+      private: void SetVisualMoved(rendering::VisualPtr _vis,
+          const bool _moved);
 
       /// \brief Qt signal when the joint creation process has ended.
       Q_SIGNALS: void JointAdded();
@@ -359,6 +429,17 @@ namespace gazebo
       /// \brief List of all links currently in the editor. The first string is
       /// the link's fully scoped name and the second is the leaf name.
       private: std::map<std::string, std::string> linkList;
+
+      /// \brief Dialog for creating a new joint.
+      private: JointCreationDialog *jointCreationDialog;
+
+      /// \brief Pose of link currently selected to be the parent of the joint
+      /// being created, before being selected.
+      private: ignition::math::Pose3d parentLinkOriginalPose;
+
+      /// \brief Pose of link currently selected to be the child of the joint
+      /// being created, before being selected.
+      private: ignition::math::Pose3d childLinkOriginalPose;
     };
     /// \}
 
@@ -417,6 +498,9 @@ namespace gazebo
 
       /// \brief Type of joint.
       public: JointMaker::JointType type;
+
+      /// \brief Last known axes values.
+      public: std::vector<ignition::math::Vector3d> axes;
 
       /// \brief True if the joint needs update.
       public: bool dirty;